+++ frontend/src/components/Mesh/TracingDiagnose.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { TestModal } from './TestModal';
+import { TestTracingModal } from './TestTracingModal';
 import { Button, ButtonVariant } from '@patternfly/react-core';
 
 type TracingDiagnoseProps = {
@@ -16,7 +16,7 @@ export const TracingDiagnose: React.FC<TracingDiagnoseProps> = (props: TracingDi
         <Button style={{ marginLeft: '5px' }} onClick={() => setIsModalOpen(true)} variant={ButtonVariant.secondary}>
           Test Config
         </Button>
-        <TestModal
+        <TestTracingModal
           isOpen={isModalOpen}
           onClose={() => setIsModalOpen(false)}
           cluster={props.cluster}
